move classes to their own files, add username back to signup form, re #2453
